Move fill and fill_diagonal to Fill.cpp, Fill.h, and FillKernel.cu
