+++ aten/src/ATen/native/cuda/UnaryOpsKernel.cu
@@ -23,21 +23,6 @@ void bitwise_not_kernel_cuda(TensorIterator& iter) {
   }
 }
 
-template <typename scalar_t>
-void fill_kernel_impl(TensorIterator& iter, Scalar value_scalar) {
-  auto value = value_scalar.to<scalar_t>();
-  gpu_kernel(iter, [value]GPU_LAMBDA() -> scalar_t {
-    return value;
-  });
-}
-
-static void fill_kernel_cuda(TensorIterator& iter, Scalar value) {
-  AT_DISPATCH_ALL_TYPES_AND2(at::ScalarType::Bool, at::ScalarType::Half, iter.dtype(), "fill_cuda", [&]() {
-    fill_kernel_impl<scalar_t>(iter, value);
-  });
-}
-
-REGISTER_DISPATCH(fill_stub, &fill_kernel_cuda);
 REGISTER_DISPATCH(bitwise_not_stub, &bitwise_not_kernel_cuda);
 
 }}
